+++ plugins/config/parsers/decorators.cpp
@@ -87,7 +87,8 @@ class DecoratorsConfigParserPlugin : public ConfigParserPlugin {
   std::map<std::string, std::vector<std::string>> load_;
 
   /// Set of configuration sources to valid intervals.
-  std::map<std::string, std::map<size_t, std::vector<std::string>>> intervals_;
+  std::map<std::string, std::map<uint64_t, std::vector<std::string>>>
+      intervals_;
 
  public:
   /// The result set of decorations, column names and their values.
@@ -263,7 +264,7 @@ void clearDecorations(const std::string& source) {
 }
 
 void runDecorators(DecorationPoint point,
-                   size_t time,
+                   uint64_t time,
                    const std::string& source) {
   if (FLAGS_disable_decorators) {
     return;
